chore: move common mock struct
