+++ .mise.toml
@@ -1,2 +1,3 @@
 [tools]
 "cargo:typos-cli" = "latest"
+"cargo:cargo-machete" = "latest"
