+++ src/repository/did_repository.rs
@@ -5,3 +5,63 @@ pub trait DidRepository {
     async fn create_identifier(&self) -> anyhow::Result<DIDResolutionResponse>;
     async fn find_identifier(&self, did: &str) -> anyhow::Result<Option<DIDResolutionResponse>>;
 }
+
+#[cfg(test)]
+pub mod mocks {
+    use std::collections::BTreeMap;
+
+    use super::*;
+    use crate::nodex::{
+        keyring::keypair::KeyPairing,
+        sidetree::payload::{DIDDocument, DidPublicKey, MethodMetadata},
+    };
+
+    #[derive(Clone)]
+    pub struct MockDidRepository {
+        map: BTreeMap<String, KeyPairing>,
+    }
+
+    impl MockDidRepository {
+        pub fn new(map: BTreeMap<String, KeyPairing>) -> Self {
+            Self { map }
+        }
+    }
+
+    #[async_trait::async_trait]
+    impl DidRepository for MockDidRepository {
+        async fn create_identifier(&self) -> anyhow::Result<DIDResolutionResponse> {
+            unimplemented!()
+        }
+        async fn find_identifier(
+            &self,
+            did: &str,
+        ) -> anyhow::Result<Option<DIDResolutionResponse>> {
+            if let Some(keyring) = self.map.get(did) {
+                let jwk = keyring.sign.to_jwk(false)?;
+
+                let response = DIDResolutionResponse {
+                    context: "https://www.w3.org/ns/did-resolution/v1".to_string(),
+                    did_document: DIDDocument {
+                        id: did.to_string(),
+                        public_key: Some(vec![DidPublicKey {
+                            id: did.to_string() + "#signingKey",
+                            controller: String::new(),
+                            r#type: "EcdsaSecp256k1VerificationKey2019".to_string(),
+                            public_key_jwk: jwk,
+                        }]),
+                        service: None,
+                        authentication: Some(vec!["signingKey".to_string()]),
+                    },
+                    method_metadata: MethodMetadata {
+                        published: true,
+                        recovery_commitment: None,
+                        update_commitment: None,
+                    },
+                };
+                Ok(Some(response))
+            } else {
+                Ok(None)
+            }
+        }
+    }
+}
