+++ src/services/did_vc.rs
@@ -119,71 +119,16 @@ impl DIDVCService {
 
 #[cfg(test)]
 mod tests {
-    use rand::distributions::{Alphanumeric, DistString as _};
+    use std::{collections::BTreeMap, iter::FromIterator as _};
+
     use serde_json::json;
 
     use super::*;
     use crate::{
-        nodex::{
-            extension::trng::OSRandomNumberGenerator,
-            keyring::keypair::KeyPairing,
-            sidetree::payload::{DIDDocument, DIDResolutionResponse, DidPublicKey, MethodMetadata},
-        },
-        repository::did_repository::DidRepository,
+        nodex::{extension::trng::OSRandomNumberGenerator, keyring::keypair::KeyPairing},
+        repository::did_repository::mocks::MockDidRepository,
+        services::test_utils::create_random_did,
     };
-
-    struct MockDidRepository {
-        did: String,
-        keyring: KeyPairing,
-    }
-
-    impl MockDidRepository {
-        pub fn new(did: String, keyring: KeyPairing) -> Self {
-            Self { did, keyring }
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl DidRepository for MockDidRepository {
-        async fn create_identifier(&self) -> anyhow::Result<DIDResolutionResponse> {
-            let res = self.find_identifier(&self.did).await?;
-            Ok(res.unwrap())
-        }
-        async fn find_identifier(
-            &self,
-            did: &str,
-        ) -> anyhow::Result<Option<DIDResolutionResponse>> {
-            // extract from NodeX::create_identifier
-            let jwk = self.keyring.sign.to_jwk(false)?;
-
-            let response = DIDResolutionResponse {
-                context: "https://www.w3.org/ns/did-resolution/v1".to_string(),
-                did_document: DIDDocument {
-                    id: did.to_string(),
-                    public_key: Some(vec![DidPublicKey {
-                        id: did.to_string() + "#signingKey",
-                        controller: String::new(),
-                        r#type: "EcdsaSecp256k1VerificationKey2019".to_string(),
-                        public_key_jwk: jwk,
-                    }]),
-                    service: None,
-                    authentication: Some(vec!["signingKey".to_string()]),
-                },
-                method_metadata: MethodMetadata {
-                    published: true,
-                    recovery_commitment: None,
-                    update_commitment: None,
-                },
-            };
-
-            Ok(Some(response))
-        }
-    }
-
-    fn create_random_did() -> String {
-        let random_string = Alphanumeric.sample_string(&mut rand::thread_rng(), 16);
-        format!("did:nodex:test:{}", random_string)
-    }
 
     #[actix_rt::test]
     async fn test_generate_and_verify() {
@@ -192,7 +137,8 @@ mod tests {
         let trng = OSRandomNumberGenerator::default();
         let from_keyring = KeyPairing::create_keyring(&trng).unwrap();
 
-        let mock_repository = MockDidRepository::new(from_did.to_string(), from_keyring.clone());
+        let mock_repository =
+            MockDidRepository::new(BTreeMap::from_iter([(from_did.clone(), from_keyring.clone())]));
 
         let service = DIDVCService::new(mock_repository);
