+++ src/services/mod.rs
@@ -1,3 +1,13 @@
 pub mod did_vc;
 pub mod encrypted;
 pub mod types;
+
+#[cfg(test)]
+pub mod test_utils {
+    use rand::distributions::{Alphanumeric, DistString as _};
+
+    pub fn create_random_did() -> String {
+        let random_string = Alphanumeric.sample_string(&mut rand::thread_rng(), 16);
+        format!("did:nodex:test:{}", random_string)
+    }
+}
